+++ UniSense/UtilityComponents/TriggerStateSetter.cs
@@ -1,0 +1,92 @@
+using UnityEngine;
+using UnityEditor;
+
+namespace UniSense
+{
+    /// <summary>
+    /// A simple component to setup and update triggerStates from the unity editor
+    /// </summary>
+    public class TriggerStateSetter : MonoBehaviour
+    {
+        public enum TriggerType
+        {
+            Left,
+            Right,
+            Both
+        }
+
+        public TriggerType triggerToSet;
+        public DualSenseSerializableTriggerState triggerState;
+
+        void Start()
+        {
+
+        }
+
+        public void Set(DualSenseGamepadHID dualSenseGamepad = null)
+        {
+            if (dualSenseGamepad is null)
+                dualSenseGamepad = DualSenseGamepadHID.FindCurrent();
+
+            if (dualSenseGamepad is null)
+            {
+                Debug.LogWarning("Current gamepad (default value) is not a dualsense");
+                return;
+            }
+            else
+            {
+                switch (triggerToSet)
+                {
+                    case TriggerType.Left:
+                        dualSenseGamepad.SetTriggerState(triggerState, null, true);
+                        break;
+
+                    case TriggerType.Right:
+                        dualSenseGamepad.SetTriggerState(null, triggerState, true);
+                        break;
+
+                    case TriggerType.Both:
+                        dualSenseGamepad.SetTriggerState(triggerState, triggerState, true);
+                        break;
+                }
+            }
+        }
+    }
+
+#if UNITY_EDITOR
+    [CustomEditor(typeof(TriggerStateSetter))]
+    [CanEditMultipleObjects]
+    public class SetTriggerForceEditor : Editor
+    {
+        SerializedProperty triggerToSetProperty;
+        SerializedProperty triggerStateProperty;
+
+        private void OnEnable()
+        {
+            triggerToSetProperty = serializedObject.FindProperty("triggerToSet");
+            triggerStateProperty = serializedObject.FindProperty("triggerState");
+        }
+
+        public override void OnInspectorGUI()
+        {
+            serializedObject.Update();
+
+            //EditorGUILayout.PropertyField(dualSenseProperty);
+            EditorGUILayout.PropertyField(triggerToSetProperty);
+            EditorGUILayout.PropertyField(triggerStateProperty);
+
+            EditorGUI.BeginDisabledGroup(!Application.isPlaying);
+            if (GUILayout.Button(new GUIContent("Call Set(First_Dualsense)", "Must be in Play mode : the gamepad is not yet binded otherwise")))
+            {
+                foreach (TriggerStateSetter setter in targets)
+                {
+                    setter.Set(DualSenseGamepadHID.FindFirst());
+                }
+            }
+            EditorGUI.EndDisabledGroup();
+
+            serializedObject.ApplyModifiedProperties();
+        }
+    }
+#endif
+}